project com.jaspersoft.studio + table : - added new FR translations - deleted unused i18ned strings - deleted int cast in Win32ImageRenderer.java
git-svn-id: http://code.jaspersoft.com/svn/repos/jaspersoftstudio/trunk@406 6833c83e-4fd0-4c68-a3df-283c6148db30

+++ com.jaspersoft.studio/src/com/jaspersoft/studio/editor/java2d/Win32ImageRenderer.java
@@ -38,7 +38,7 @@ public class Win32ImageRenderer implements ImageRenderer {
 	  {
   		Integer intParamGC = Integer.class.cast(paramGC.handle);
   		
-		  renderImage((int) intParamGC.intValue(), xDest, yDest, width, height, 0, 0, data, imgWidth, imgHeight);
+		  renderImage(intParamGC.intValue(), xDest, yDest, width, height, 0, 0, data, imgWidth, imgHeight);
 	  }
 
 	  static {
